requirements.txt
redis~=4.5.1
numpy~=1.24.2
scipy~=1.10.1
grpcio~=1.51.3
grpcio-tools~=1.51.3
python-dotenv~=1.0.0
